split plan blueprint generator into a separate file
